requirements.txt
fitz==0.0.1.dev2
Flask==2.2.2
Flask_Bcrypt==1.0.1
openai==0.28.0
plotly==5.9.0
python-dotenv==1.0.0
Werkzeug==2.2.3
